add animation to logo
